product_batch_processor/requirements.txt
# Core dependencies
aiohttp>=3.9.0
asyncio>=3.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web crawling
firecrawl-py>=0.0.16

# Google Cloud
google-cloud-aiplatform>=1.38.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
vertexai>=1.38.0

# Image processing
Pillow>=10.0.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# CLI and progress
click>=8.1.0
argparse